+++ packages/ui/src/components/StravaActivityImageGenerator/Content/Image/downloadBase64Image/index.ts
@@ -1,0 +1,1 @@
+export { default } from './downloadBase64Image';