+++ packages/ui/src/pages/ActivitiesPage/Activities/ActivitiesList/ActivitiesList.tsx
@@ -1,22 +1,22 @@
 import { Card, Button, Text, Grid, Spacer } from '@geist-ui/core';
 import { Activity as ActivityIcon, Navigation, Clock, TrendingUp, Zap } from '@geist-ui/icons';
+import { StravaActivity } from '@pace/strava-api';
 
-import { Activity } from '../../../../api/strava';
 import formatActivityType from './formatActivityType';
 
 interface ActivitiesListProps {
-  activities: Activity[];
-  generateImage: (activityId: number) => Promise<void>;
+  activities: StravaActivity[];
+  onGenerateImage: (activityId: string) => void;
 }
 
 /**
  * Activities list view.
  * @param {ActivitiesListProps} props - Component props.
- * @param {Activity[]} props.activities - List of activities to display.
- * @param {Function} props.generateImage - Function to generate image for an activity.
+ * @param {StravaActivity[]} props.activities - List of activities to display.
+ * @param {Function} props.onGenerateImage - Function to open the activity image generation view.
  * @returns {JSX.Element} Activities list view.
  */
-const ActivitiesList = ({ activities, generateImage }: ActivitiesListProps) =>
+const ActivitiesList = ({ activities, onGenerateImage }: ActivitiesListProps) =>
   activities.map((activity) => (
     <Grid xs={24} sm={12} md={8} key={activity.id}>
       <Card width="100%" hoverable>
@@ -40,17 +40,17 @@ const ActivitiesList = ({ activities, generateImage }: ActivitiesListProps) =>
                 <ActivityIcon size={14} /> {formatActivityType(activity.type)}
               </Text>
             )}
-            {activity.distance > 0 && (
+            {(activity?.distance ?? 0) > 0 && (
               <Text small>
-                <Navigation size={14} /> {(activity.distance / 1000).toFixed(2)} km
+                <Navigation size={14} /> {((activity.distance ?? 0) / 1000).toFixed(2)} km
               </Text>
             )}
-            {activity.moving_time > 0 && (
+            {(activity?.moving_time ?? 0) > 0 && (
               <Text small>
-                <Clock size={14} /> {Math.floor(activity.moving_time / 60)} min
+                <Clock size={14} /> {Math.floor((activity.moving_time ?? 0) / 60)} min
               </Text>
             )}
-            {activity.total_elevation_gain != null && activity.total_elevation_gain > 0 && (
+            {(activity?.total_elevation_gain ?? 0) > 0 && (
               <Text small>
                 <TrendingUp size={14} /> {activity.total_elevation_gain} m
               </Text>
@@ -64,7 +64,7 @@ const ActivitiesList = ({ activities, generateImage }: ActivitiesListProps) =>
             scale={0.8}
             icon={<Zap />}
             onClick={() => {
-              generateImage(activity.id).catch(console.error);
+              onGenerateImage(String(activity.id));
             }}
             placeholder="Generate Image"
             onPointerEnterCapture={() => undefined}